feat: revoke allowance

+++ contracts/protocol/proxies/RigoblockPoolProxyFactory.sol
@@ -162,8 +162,7 @@ contract RigoblockPoolProxyFactory is IRigoblockPoolProxyFactory {
         view
         returns (bool)
     {
-        // TODO: test uin256 vs uint32
-        uint32 size;
+        uint256 size;
         assembly {
             size := extcodesize(_target)
         }
